review fixes 1
